reports: add JasperReport sale order template
Add a template for sale order using Jasper Reports engine. This fixes a
bunch of BIRT-related bugs and enhance display. The report is not
perfect as it uses JasperUtils to run a query where a list + return
values will perfectly fit. This is due to hesitation during design and
should be fixed in a future update.